Schema2 parquet import storage created.
Impala2 import storage now inherits from Schema2 import storage.
